renamed CodeEditor to CodeEditorComponent
