+++ playground/src/components/CodeEditorComponent.tsx
