+++ playground/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Navigation } from './Navigation';
-export { default as CodeEditor } from './CodeEditor';
+export { default as CodeEditorComponent } from './CodeEditorComponent';
 export { default as FetchSchema } from './FetchSchema';
 export { default as Tabs } from './Tabs';
 export { default as Tab } from './Tab';
